dynamically rendering key sound list
