+++ Metro/app.js
@@ -13,6 +13,7 @@ import { loadUser } from '../Common/load-user.js';
 import { changeTheme } from '../Common/change-theme.js';
 //import { SoundBoard } from '../utils/make-sound-board.js'; //class
 import { whiteKeysColorChange, blackKeysColorChange } from './color-change.js';
+import { generateKeySoundListItem } from '../utils/generateKeySoundListItem.js';
 
 const saveSound = document.getElementById('save-sound');
 //let soundBoard = new SoundBoard(soundBoards[0]);
@@ -29,11 +30,13 @@ createHeader();
 loadTheme();
 whiteKeysColorChange();
 blackKeysColorChange();
+generateKeySoundListItem(soundBoards);
 
 selectMenu.addEventListener('input', changeTheme);
 
 const metroSoundForm = document.getElementById('metronome-sound');
 metroSoundForm.appendChild(generateMetroSoundList(metroSounds));
+
 
 const start = document.getElementById('start');
 start.addEventListener('click', () => {
