+++ Metro/index.html
@@ -121,7 +121,7 @@
 
             <form>Keyboard Sounds 
                 <select name="soundbank" id="select-soundbank">
-                    <option value = "./data/soundboard.js">SoundBoard</option>
+                    <!-- <option value = "./data/soundboard.js">SoundBoard</option> -->
                 </select>
                 
                 <label for="color-scheme">Color Scheme:</label>
